+++ mb-url.el
@@ -102,6 +102,9 @@
 ;; (setq mb-url-http-backend 'mb-url-http-curl)
 ;; ```
 ;;
+;; NOTE: All clients are required to support automatically decompressing
+;; response bodies.
+;;
 ;; #### [cURL][]
 ;;
 ;; [cURL]: http://curl.haxx.se/